+++ ARRRtomatic_diff/auto_diff.py
@@ -874,7 +874,7 @@ class AutoDiffVector:
         return AutoDiffVector.combine(self, other, lambda x,y: x // y).get_values()
 
     def __mod__(self, other):
-        return AutoDiffVector.combine(self, other, lambda x,y: x & y).get_values()
+        return AutoDiffVector.combine(self, other, lambda x,y: x % y).get_values()
 
     def __lshift__(self, other):
         return AutoDiffVector.combine(self, other, lambda x,y: x << y).get_values()
@@ -895,7 +895,7 @@ class AutoDiffVector:
         return AutoDiffVector.combine(other, self, lambda x,y: x // y).get_values()
 
     def __rmod__(self, other):
-        return AutoDiffVector.combine(other, self, lambda x,y: x & y).get_values()
+        return AutoDiffVector.combine(other, self, lambda x,y: x % y).get_values()
 
     def __rlshift__(self, other):
         return AutoDiffVector.combine(other, self, lambda x,y: x << y).get_values()
